+++ test/dict-helpers.js
@@ -180,8 +180,8 @@ function getFilledFullMmtfDict(){
         sequenceIndexList: new Uint8Array( MmtfUtils.makeInt32Buffer( [ 0, 1 ] ) ),
 
         // chains
-        chainIdList: new Uint8Array( [ 0, 0, 68, 65 ] ),
-        chainNameList: new Uint8Array( [ 0, 0, 0, 66 ] ),
+        chainIdList: new Uint8Array( [ 68, 65, 0, 0 ] ),
+        chainNameList: new Uint8Array( [ 66, 0, 0, 0 ] ),
         groupsPerChain: [ 1 ],
 
         // models
@@ -226,7 +226,7 @@ function getFilledRequiredMmtfDict(){
         groupTypeList: new Uint8Array( MmtfUtils.makeInt32Buffer( [ 0 ] ) ),
 
         // chains
-        chainIdList: new Uint8Array( [ 0, 65, 0, 0 ] ),
+        chainIdList: new Uint8Array( [ 65, 0, 0, 0 ] ),
         groupsPerChain: [ 1 ],
 
         // models